+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.40b11"  # see https://peps.python.org/pep-0440/
-    AppDateStr = "2025-06-25"
+    AppVersionStr = "0.40b12"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2025-07-02"